+++ pype/lib/lib_old.py
@@ -327,39 +327,6 @@ def get_last_version_from_path(path_dir, filter):
         return None
 
 
-def ffprobe_streams(path_to_file, logger=None):
-    """Load streams from entered filepath via ffprobe."""
-    if not logger:
-        logger = log
-    logger.info(
-        "Getting information about input \"{}\".".format(path_to_file)
-    )
-    args = [
-        "\"{}\"".format(get_ffmpeg_tool_path("ffprobe")),
-        "-v quiet",
-        "-print_format json",
-        "-show_format",
-        "-show_streams",
-        "\"{}\"".format(path_to_file)
-    ]
-    command = " ".join(args)
-    logger.debug("FFprobe command: \"{}\"".format(command))
-    popen = subprocess.Popen(
-        command,
-        shell=True,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE
-    )
-
-    popen_stdout, popen_stderr = popen.communicate()
-    if popen_stdout:
-        logger.debug("ffprobe stdout: {}".format(popen_stdout))
-
-    if popen_stderr:
-        logger.debug("ffprobe stderr: {}".format(popen_stderr))
-    return json.loads(popen_stdout)["streams"]
-
-
 def source_hash(filepath, *args):
     """Generate simple identifier for a source file.
     This is used to identify whether a source file has previously been
